# go/model_bookshelf.go
/*
 * Book Blog API
 *
 * This is a blog about books.
 *
 * API version: 0.0.7
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type Bookshelf struct {

	Num int32 `json:"num"`

	Bookshelf []BookshelfBookshelf `json:"bookshelf"`
}