# go/model_review.go
/*
 * Book Blog API
 *
 * This is a blog about books.
 *
 * API version: 0.0.7
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type Review struct {

	ID int32 `json:"ID"`

	Content string `json:"Content,omitempty"`

	Autor string `json:"autor,omitempty"`
}